requirements.txt
astroid==2.4.2
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
